// src/com/siigo/certification/front/models/CustomerModel.ts
export interface CustomerModel {
    tipoPersona: string;
    tipoIdentificacion: string;
    numeroIdentificacion: string;
    dv: string;
    codSucursal: string;
    nombres: string;
    apellidos: string;
    razonSocial: string;
    ciudad: string;
    direccion: string;
    indicativo: string;
    telefono: string;
    extension: string;
    correo: string;
    tipoIva: string;
    indicativoFactura: string;
    telefonoFactura: string;
    codigoPostal: string;
    nombreContacto: string;
    apellidoContacto: string;
    emailContacto: string;
    cargo: string;
    indicativoContacto: string;
    telefonoContacto: string;
}          